requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
torch==2.10.0
transformers==5.0.0
faiss-cpu==1.13.2
numpy==2.4.1
pillow==12.1.0
datasets==4.5.0
huggingface_hub==1.3.4
python-multipart==0.0.22
tokenizers==0.22.2
safetensors==0.7.0
requests==2.32.5
tqdm==4.67.1
